[#30] Feat : cardlist context, reducer, action 분리
- cardList 내부의 상태관리 로직은 제거
