+++ FE/src/actions/actions.js
@@ -1,0 +1,6 @@
+const fetchActions = {
+  FETCH_SUCCESS: "FETCH_SUCCESS",
+  FETCH_ERROR: "FETCH_ERROR",
+};
+
+export { fetchActions };
